Changed component based library flow so the query comes directly from the state (and not from the router), fixed pagination, added sql script for generating a lot of data and changed distance from integer (meters) to (double) km.

+++ src/main/kotlin/pt/isel/ls/sports/database/sections/routes/RoutesDB.kt
@@ -18,7 +18,7 @@ interface RoutesDB {
      *
      * @return the route's unique identifier
      */
-    fun createNewRoute(conn: ConnectionDB, startLocation: String, endLocation: String, distance: Int, uid: Int): Int
+    fun createNewRoute(conn: ConnectionDB, startLocation: String, endLocation: String, distance: Double, uid: Int): Int
 
     /**
      * Get the details of a route.
